Bug 35202: Table settings should apply to multiple subscriptions in the OPAC
This patch updates the bibliographic detail page in the OPAC so that if
there are multiple subscriptions attached to one record, each table of
issues under the subscriptions tab adheres to the settings defined in
Table Settings.

To test you should have at least two subscriptions attached to one
record, and at least one issue of each subscription.

- Apply the patch and view the record in the OPAC.
- Under the "Subscriptions" tab there should be separate tables for each
  subscription, each of them a sortable DataTable.
- In the staff interface, go to Administration -> Table settings -> OPAC
  -> subscriptionst
  - Select some columns to be hidden by default.
- Return to the OPAC and confirm that each subscription table now
  reflects the change you made in table settings.

Sponsored-by: Athens County Public Libraries
Signed-off-by: David Nind <david@davidnind.com>
Signed-off-by: Ray Delahunty <lib-systems@arts.ac.uk>

Signed-off-by: Marcel de Rooy <m.de.rooy@rijksmuseum.nl>
Signed-off-by: Katrin Fischer <katrin.fischer@bsz-bw.de>