Bug 39083: Fix title sorting on list of tagged titles
This patch adds the "anti-the" class to the title column on the tagged
titles page. This allows DataTables to ignore articles when sorting.

To test you should add the same tag to multiple titles.

In the staff interface, go to Tools -> Tags, and click a term which has
been added to multiple titles.

In the list of titles, test title sorting. Articles like "a", "an", and
"the" should be ignored when sorting.

Sponsored-by: Athens County Public Libraries
Signed-off-by: Phil Ringnalda <phil@chetcolibrary.org>
Signed-off-by: Katrin Fischer <katrin.fischer@bsz-bw.de>