Bug 38664: Revert changse made to t/db_dependent/misc/translator/sample.tt
We don't want to comment the whole block or it won't be parsed by the
translator tool.

This has been done initially to make the prettier plugin happy but it
didn't handle TT comments correctly. Now with version 2.0.2 it should be
better.

This patch fixes t/db_dependent/misc/translator/xgettext-tt2.t