Bug 39124: In lists dropdown, the option "view all" is always displayed
In the code of the OPAC, the "Lists" dropdown menu is supposed to
display max. 10 lists and display a button "view all" if there are more
than 10 lists.

The button "view all" is always displayed and it is weird when there
are few lists because "Public lists" and "Your lists" do the same.

Test plan:
1 - On an OPAC with less than 10 public lists and 10 private lists,
    navigate to /cgi-bin/koha/opac-shelves.pl, click on the dropdown see
    there is a view all button for both types of lists
2 - APPLY PATCH
3 - Repeat 1, notice the buttons disappeared
4 - Create lists so there are >10 public and >10 private lists
5 - Repeat 1, notice the buttons reappeared

Signed-off-by: Owen Leonard <oleonard@myacpl.org>

Signed-off-by: Marcel de Rooy <m.de.rooy@rijksmuseum.nl>
Signed-off-by: Katrin Fischer <katrin.fischer@bsz-bw.de>
(cherry picked from commit 88a292af36a142117b4b81b005e134992e272df7)
Signed-off-by: Baptiste Wojtkowski <baptiste.wojtkowski@biblibre.com>