Bug 39145: Apply new pref values to ->disown_or_delete
Test plan:
Set ListOwnershipUponPatronDeletion to 'change owner of
these lists'.
Create a public list1 and a private list2 for patron A.
Enable sharing lists and share list2 with another patron.
Remove patron A.
Check that list1 and list2 are transferred.
Set  ListOwnershipUponPatronDeletion to 'change owner of
public lists, delete shared lists'.
Create a public list3 and a private list4 for patron B.
Share list4 with another patron.
Remove patron B.
Check that list3 is transferred and list4 is deleted.

Signed-off-by: Marcel de Rooy <m.de.rooy@rijksmuseum.nl>
Signed-off-by: David Nind <david@davidnind.com>
Signed-off-by: Martin Renvoize <martin.renvoize@openfifth.co.uk>
Signed-off-by: Lucas Gass <lucas@bywatersolutions.com>