Bug 32051: Rename Library to Contact for one public library
Note:
I am using $rs in the script to prevent searching branches again
unneeded. And pass public_count to opac-library.tt to prevent
another call of Branches.all (for bread crumbs), like masthead
already does btw.

Test plan:
[1] Have two (or more) public libraries.
Verify that you see Libraries on the OPAC home page and that when
clicking Libraries, you first have the library list. Are the bread
crumbs as expected?
Click a library on that list. Do you get its details? Are the bread
crumbs as expected?

[2] Now have only one public library.
Verify that you see Contact on the home page.
And that when clicking Contact you go directly to the library details
without the intermediate library list. Are the bread crumbs still
as expected (with Contact instead of Libraries)?

Signed-off-by: Marcel de Rooy <m.de.rooy@rijksmuseum.nl>
Signed-off-by: Owen Leonard <oleonard@myacpl.org>
Signed-off-by: Paul Derscheid <paul.derscheid@lmscloud.de>
Signed-off-by: Katrin Fischer <katrin.fischer@bsz-bw.de>