Bug 39217: Fix ugly error on wrong password in SCO login
This patch updates the self checkout page so that it uses the same
message as the regular OPAC in response to a user trying to log in with
incorrect credentials.

To test, apply the patch enable the self checkout module if necessary.

- Set the SelfCheckoutByLogin system preference to "cardnumber"
- Test the process of logging in as a patron to self checkout:
  Submitting an invalid library card number should trigger an error.
- Set the SelfCheckoutByLogin system preference to "username and
  password."
- Test this login form by submitting a valid card number or username and
  an invalid password. You should see the same error message.
- NOTE: If you submit an invalid card number or username you will get no
  message at all, which is a bug in and of itself.

QA note: I think the qa tool warning is an error in the tidy check.

Sponsored-by: Athens County Public Libraries
Signed-off-by: David Nind <david@davidnind.com>
Signed-off-by: Lucas Gass <lucas@bywatersolutions.com>
Signed-off-by: Katrin Fischer <katrin.fischer@bsz-bw.de>