Bug 39299: Fix cataloguing plugins in serial item editor
Some CSS classes were missing, causing the corresponding event listener
to not be called.

Adding the missing CSS classes caused the tag editor button to be moved
to the left of the input, because of 'float: left' for .tag_editor
It seems that float:left is useless now, at least in cataloguing where
we use 'display: flex' for those elements, so this patch removes it.

Test plan:
1. Set 'cn_browser.pl' as a plugin for one of the item subfields (e.g.
   952$o)
2. Go to the serial edition page and click on "Add item". This should
   show an item form
3. Click on the tag editor button, the cn_browser.pl popup should appear

Signed-off-by: Martin Renvoize <martin.renvoize@ptfs-europe.com>
Signed-off-by: Katrin Fischer <katrin.fischer@bsz-bw.de>