Bug 39293: Remove box around subfield tag in basic editor
Prior to the Bootstrap 5 update, the feature of editing subfield codes in the
basic editor was kept subtle by putting the code in an <input>, but removing
the border around the input, so it required either knowing the feature exists
or accidentally clicking on or around the single letter code and seeing it
turn into an editable input.

This patch removes the border again, because when people accidentally see the
feature, they are sometimes not pleased to see it.

Test plan:
1. Without the patch, open the cataloging basic editor and switch to the "1"
   tab with the 100 fields.
2. Note that the codes for the subfields are clearly in editable inputs.
3. Apply patch, reload the editor page.
4. Note that now the codes look like just plain text, until you actually
   click on or very near the letter, when they become visible as inputs
   by having the focus border.

Sponsored-by: Chetco Community Public Library
Signed-off-by: Andrew Fuerste Henry <andrew@bywatersolutions.com>
Signed-off-by: Baptiste Wojtkowski <baptiste.wojtkowski@biblibre.com>
Signed-off-by: Katrin Fischer <katrin.fischer@bsz-bw.de>