Bug 39350: (bug 38664 follow-up) Restore language selector for Vue.js module
ERM and Preservation modules have lost the language selector during the
tidy work.
There has been a closing div added to the "home" files, which broke the
HTML structure.

Test plan:
Have more than one language selected for StaffInterfaceLanguages.
Go to the ERM and Preservation module and confirm that with this patch
applied the language selector menu is displayed.

Signed-off-by: David Nind <david@davidnind.com>
Signed-off-by: Martin Renvoize <martin.renvoize@openfifth.co.uk>
Signed-off-by: Katrin Fischer <katrin.fischer@bsz-bw.de>