Bug 39354: Remove unintended Bootstrap 5 change to scroll-behavior
This patch updates the "variables" SCSS file in the OPAC and staff
interface in order to disable the "smooth scroll" behavior introduced in
Bootstrap 5. This was an unexpected and unintentional change.

The patch also corrects another unexpected side-effect of the BS5
upgrade: A change to the "cursor" property of elements with a
".placeholder" class. This class is used by the multiple-select jQuery
plugin found in some system preference controls.

To test, apply the patch and rebuild the staff and OPAC CSS.

- Clear your browser cache if necessary.
- In the staff interface, go to Administration -> System preferences ->
  Circulation, and click "Patron restrictions" in the left-hand sidebar.
  The page should jump immediately to that section without any scrolling
  effect.
- Go to the "Article requests" section of that page. Hover your cursor
  over a dropdown which doesn't have any option selected. Your cursor
  should not change to a "waiting" cursor.

Signed-off-by: David Nind <david@davidnind.com>

Signed-off-by: Jonathan Druart <jonathan.druart@bugs.koha-community.org>
Signed-off-by: Katrin Fischer <katrin.fischer@bsz-bw.de>
(cherry picked from commit 09c79110a11e5d336734357328c858b3245c271b)
Signed-off-by: Fridolin Somers <fridolin.somers@biblibre.com>