Bug 39395: Fix #wrapper closing tag in sco
The module was displayed at the bottom of the screen.
It's caused by the closing tag of #wrapper that has been added by bug
38664 inside the include file.

It is wrong, the div.main should be within #wrapper.

The correct fix is to move #wrapper out of the include file, so that the
HTML structure is correct (and the prettier plugin is happy) in the .tt
files

Test plan:
Confirm the bug
Apply this patch
Go to the SCO module and confirm that the module is correctly placed at
the top of the screen

Please review this patch using `git show -w HEAD`

Signed-off-by: David Nind <david@davidnind.com>
Signed-off-by: Lucas Gass <lucas@bywatersolutions.com>
Signed-off-by: Katrin Fischer <katrin.fischer@bsz-bw.de>